Implement a IsometricStarPolygon class
