+++ src/@classes/abstract/IsometricStarPolygonAbstract/index.ts
@@ -1,0 +1,2 @@
+export { IsometricStarPolygonAbstract } from './IsometricStarPolygonAbstract';
+export * from './types';
