+++ src/@classes/public/IsometricStarPolygon/index.ts
@@ -1,0 +1,2 @@
+export { IsometricStarPolygon } from './IsometricStarPolygon';
+export { IsometricStarPolygonProps } from './types';
