+++ src/@classes/public/IsometricStarPolygon/types.ts
@@ -1,0 +1,3 @@
+import { IsometricStarPolygonAbstractProps } from '@classes/abstract/IsometricStarPolygonAbstract';
+
+export type IsometricStarPolygonProps = IsometricStarPolygonAbstractProps;
